TransactionFactory.fromDb => TransactionWithXDRFactory
